### 2022-04-10 #4307 @arfon
@mlampros – many thanks for your submission to JOSS. A few issues to deal with before we get started looking for a handling editor:

- Please rename your paper to `paper.md` and remove the `citation_author` string in the Markdown file. I've opened a PR that does both of these for you here: https://github.com/mlampros/IceSat2R/pull/2
- Please ensure your repository contains a plain text OSI-approved license file called `LICENSE`

Finally, as this package looks to be an API wrapper which is often not in scope for JOSS, I'm going to ask the broader editorial team for their input on whether we consider this submission in scope for us. Scope reviews generally take a week or two to complete.
