+++ .github/workflows/joss_paper.yml
@@ -19,7 +19,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: inst/joss_paper/joss_paper.md
+          paper-path: inst/joss_paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -27,4 +27,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: inst/joss_paper/joss_paper.pdf
+          path: inst/joss_paper/paper.pdf
