### 2022-04-10 #4307 @mlampros
@arfon thank you for the PR (just merged). I also added the LICENSE.md file. I'll wait for the scope review.
