### 2022-04-25 #4342 @editorialbot
[REVIEW]: ICESat-2 Altimeter Data using R
**Submitting author:** @mlampros (<a href="http://orcid.org/0000-0002-8024-1546">Lampros Mouselimis</a>)
**Repository:** https://github.com/mlampros/IceSat2R
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
