### 2022-04-26 #4342 @Jean-Romain
In the first part my review is following point-by-point the authors/reviewers guidelines by citing the guidelines and providing my associated comments to discuss if the paper and software fit the guidelines. In the second part my review analyses the code and the features of the package more in depth.

# JOSS guidelines?

> JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library, or similar. While useful for many areas of research, pre-trained machine learning models and notebooks are not in-scope for JOSS.

The `IceSat2R` package is a package that essentially builds string queries to download data from a remote server using the public API provided by the service. In this sense it does not solve complex problem, it does not extract knowledge from large data sets, it does not offer a mathematical library. While useful, it does not meet the criterion mentioned above.

> As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual. [...]

The package essentially build strings to send queries and download data from a remote server. The core of the `IceSat2R` is made of approximately 1000 lines of R code with a large portion of the code doing nothing but handling user's input errors (which is normal and expected in R code) and the programming style significantly increases the number of lines. Even considering the time spent at writing the documentation this definitively cannot be considered as 3 months of work for a single person.

> Some factors that may be considered by editors and reviewers when judging effort include:
>
> - Age of software (is this a well-established software project) / length of commit history.
> - Number of commits.
> - Number of authors.
> - Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
> - Whether the software has already been cited in academic papers.
> - Whether the software is sufficiently useful that it is likely to be cited by your peer group.

The software is very young (first release 2 months ago) and definitively not well-established. As mentioned above the number of line of code is approximately 1000. I don't know if the software as already been cited but considering it is two months old I think it is temporarily impossible that `IceSat2R` has citations. Considering I'm not working with ICESat data I can't tell if the software is useful but in my own daily work I regularly appreciate that people spent time at proving easy to use tools to get access to data. The question is to know if the software is *easy* and this will be discussed later.

> In addition, JOSS requires that software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards for the programming language being used (e.g., Python, R), and designed for maintainable extension (not one-off modifications of existing tools). “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.

I'd say that the software is not feature complete and does not meet R standards and I will explain why later when analyzing the code and the features. I think this package is a thin API client. This is not bad in the absolute but a little light for a publication.

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

I wanted to comment that the paper was particularly short and empty then I read this guideline. Excluding source code the paper contains approximately 50 lines of text which means something like 700 words. It fits with the guideline but I think that all the bullets mentioned in the guidelines and not fulfilled

- [x] A list of the authors of the software and their affiliations, using the correct format (see the example below).
- [x] A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- [ ] A Statement of Need section that clearly illustrates the research purpose of the software.
- [ ] A list of key references, including to other software addressing related needs. Note that the references should include full names of venues, e.g., journals and conferences, not abbreviations only understood in the context of a specific discipline.
- [ ] Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
- [x] Acknowledgement of any financial support.

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

No. For example the man page of `available_RGTs()` says *Reference Ground Tracks (RGTs)*. That's it! This does not help me to understand what the function does. The argument `technical_specs_url` is virtually not documented because the only way to understand it is to already know what it does. The man page of `time_specific_orbits()` says *Time Specific Orbits*. The man page of `getTrack()` says *Get the ICESAT-2 Tracks*. I don't know what to do with that. This is far to be enough to reach R standards. One of the strength of R is the documentation. I consider the package not documented.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

No. The CRAN checks all documentation examples. In `IceSat2R` all the examples are wrapped into `\dontrun` (or worse `if (FALSE) {}` which means that the CRAN cannot test a single example of the documentation. The unit tests are all designed to only test error handling, not functionalities. So, in practice, for what I've seen not a single line of code coding for actual features is actually tested. And the software does not work as we will see later.

>  Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

I won't put any judgement on English, it is not my tongue but the R code content must be revised. See my comments later.

> Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

No. The paper contains 3 references including a reference to the package itself (so 2 references). I'm not in ICESat stuff but this definitively looks light. There is no other API in other languages/software for ICEsat data?

# In depth review of the package

First of all I did not run the code shown in the paper because it is written in such a way that it is not in copy/pastable format. I can comment on it anyway. I think the code is unnecessarily complex. According to what I understood I think it should not be more than 10 or 20 lines. Why it is so long and complex can be explained by the fact the package is not feature-complete and does not meet R community standards. It is going to be hard to explain without being able to copy paste code chunks and without being able to refer to a specific line (code lines are not numbered) but:

1. The minimal reproducible examples (MRE) provided in the paper requires to loads 12 third party packages. For such MRE I'm not expecting to need more that 1 to 3 packages i.e `IceSat2r` of course, `sf` of course and that's it.
2. `rds` data should belong in `data/` and be loaded with `data(goms_himal)`.
3. Why so much `sf` data pre-processing? I'm expecting `IceSat2r` to accept `sf` objects without the need to make tedious extraction of wkt string in loop before to call `IceSat2r` functions. This point really goes into *community standards*. The strength of R is that everybody can use extremely easily very complex and advanced stuff. Here author put an extra layer of complexity that can easily goes into the `IceSat2r` functions.
4. Why are they so much `for` loop? First, all for-loops should be replaced by `apply` family functions. But yet it is unnecessarily complex. The `IceSat2r` functions should be natively vectorized and users should never need to loop themselves.
5. Why is it so complex to call a function with a bounding box as input? We need to call 4 times `as.numeric(bbox["side"])`. `IceSat2r` functions should accept `bbox` objects from `sf` natively and optionally (but recommended) `terra` and `raster` `extent` objects.

Overall the code presented in the article is tedious to read and understand but could be very simple and smooth by vectorizing the functions and by doing all the pre-processing job **internally**. In R we are expecting all of this to be done by the package not by the user. This is why I said earlier that the package is not feature-complete.

Then I tested vignette 1 but the code bugged after line 9. I was not able to go further

```r
avail_cycles = available_RGTs(only_cycle_names = TRUE)
#> Error in seq.int(0, to0 - from, by) : 'to' must be a finite integer
```

Then I tested vignette 2. Same error. As a conclusion **I was not able to test the package at all**. I see in the README there is a shiny app but can't find easily how to open it.

# Conclusions

1. Document the package
2. Vectorize the functions internally
3. Accept high level object as input `sf::data.frame`,`sf::bbox`, `raster::Extent`, `terra::ext`
4. Add real unit tests. I understand that the package download data and if the server is down the package tests may fail on CRAN. This is not a fatality and there are methods to cope with that cleanly.
5. Simplify your tutorial code (this should be straightforward after dealing with point 2 and 3)
6. Debug your code. I was not able to run a single example using CRAN version. If point 4 was respected this would have never happened.
