### 2022-04-26 #4342 @mlampros
Dear @Jean-Romain thank you for reviewing the code of the IceSat2R package and for your time spent to go through the code. I'd like to ask if I'm allowed to comment on your review?
