### 2022-04-26 #4342 @Jean-Romain
I'm ok with that and for what I've understood of the review policy I think it is encouraged
