### 2022-04-26 #4342 @hugoledoux
It is encouraged indeed.

Thanks @Jean-Romain for this quick (a record I think!) and thorough review.
