### 2022-04-27 #4342 @mlampros
> > The IceSat2R package is a package that essentially builds string queries to download data from a remote server using the public API provided by the service. In this sense it does not solve complex problem, it does not extract knowledge from large data sets, it does not offer a mathematical library. While useful, it does not meet the criterion mentioned above.

> solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering);

That's true the current state of the IceSat2R package does not solve complex modelling problems

> offers a mathematical library

The IceSat2R package is not a mathematical library

> supports the functioning of research instruments or the execution of research experiments;

Let's see from the perspective of an R programming user, who might be a potential scientist and involved in remote sensing but does not have the time and knowledge to access within a time interval (days or week) IceSat-2 data from the [OpenAltimetry API ](https://openaltimetry.org/). How can the IceSat2R package help this person perform potential research experiments?

- there are three vignettes ([one](https://mlampros.github.io/IceSat2R/articles/IceSat-2_Mission_Orbits_HTML.html), [two](https://mlampros.github.io/IceSat2R/articles/IceSat-2_Atlas_products_HTML.html), [three](https://mlampros.github.io/IceSat2R/articles/IceSat-2_Virtual_File_System_Orbits_HTML.html)) explaining the functionality in detail and moreover how an R user can combine IceSat-2 data with other external data sources (such as elevation data). Additionally, the third vignette shows how the user can decrease the computation time by using virtual files rather than downloading all Reference Ground Tracks.
- each (exported) function of the R package includes an [example section and documentation](https://mlampros.github.io/IceSat2R/reference/index.html) (even functions that could be used as internal, such as "[available_RGTs()](https://mlampros.github.io/IceSat2R/reference/available_RGTs.html)")
- besides the interface to the OpenAltimetry API there is a connection to the [NSIDC](https://icesat-2.gsfc.nasa.gov/science/specs) services to download the latest Reference Ground Tracks (and cycles) that are required to verify and validate the output tracks of the OpenAltimetry API (these are updated regularly using a [Github Action](https://github.com/mlampros/IceSat2R/blob/master/.github/workflows/icesat2_tech_specs.yaml)). [Figure 2](https://github.com/mlampros/IceSat2R/blob/master/inst/joss_paper/paper.pdf) of my submitted paper shows how the OpenAltimetry and NSIDC data can be used for this purpose.

> extracts knowledge from large data sets

The [official web page](https://openaltimetry.org/) mentions _"OpenAltimetry is a cyberinfrastructure platform for discovery, access, and visualization of data from NASA’s ICESat and ICESat-2 missions"_
The R user can take advantage of the OpenAltimetry API from within R to download IceSat-2 data in a manageable way. It's not feasible to download vast amounts of data without using queries and this is the purpose of the API's.
Once the user has access to data of an Area of Interest, he/she can use the IceSat2R package to combine IceSat-2 data with other external data sources and that way to extract valuable information (I explained it using only one external data source in my [second vignette](https://mlampros.github.io/IceSat2R/articles/IceSat-2_Atlas_products_HTML.html))
