### 2022-04-27 #4342 @mlampros
> Even considering the time spent at writing the documentation this definitively cannot be considered as 3 months of work for a single person.

I've started working on the IceSat2R package in middle of January 2022 (after signing the contract with the R consortium) and I finished it (including uploading the code and creating the various Github Actions) in the first weeks of March 2022.
I have to mention, though, that I was interested in IceSat-2 data before starting writing the code and towards the end of 2021 I spent 2 to 3 weeks to read papers, articles, tutorials and existing code.
