### 2022-04-27 #4342 @mlampros
> No. For example the man page of available_RGTs() only says Reference Ground Tracks (RGTs). That's it! This does not help me to understand what the function does. The argument technical_specs_url is virtually not documented because the only way to understand it is to already know what it does. The man page of time_specific_orbits() only says Time Specific Orbits. The man page of getTrack() only says Get the ICESAT-2 Tracks. I don't know what to do with that. This is far to be enough to reach R standards. One of the strength of R is the documentation. I consider the package not documented.

The functions,
