### 2022-04-27 #4342 @mlampros
> The CRAN checks all documentation examples. In IceSat2R all the examples are wrapped into \dontrun (or worse if (FALSE) {} which means that the CRAN cannot test a single example of the documentation. The unit tests are all designed to only test error handling, not functionalities

The reason that I didn't include all tests (but only tests related to error handling) was because the majority of the functions make calls to the OpenAltimetry API. I've already submitted many R packages to CRAN and based on my experience this might cause errors not only during the first submission of an R package but also during regular checks of CRAN
