### 2022-04-27 #4342 @mlampros
>  In addition the vignettes that are expected to be generated and tested by the CRAN are actually hard coded and not a single line of code is actually ran. So, in practice, for what I've seen not a single line of code coding for real features is actually tested.

This has an explanation too. There is always a trade off when submitting a package to CRAN.
Initially I intended to submit the vignettes in .html format where the code that didn't make any calls to the OpenAltimetry API would be auto-generated by CRAN (this allows a package maintainer to show potential visualizations in the true format - a leaflet map where an R user can zoom in and out)
Then I realized that this approach would create a compressed _.tar.gz_ file of approximately 13.5 MB which is way bigger than the allowed uploaded file size by CRAN (5 MB). Thus, I had to either exclude vignettes (from the 3 I would have to submit only 1). I decided to include only the necessary plots and visualizations in reduced size and to not auto generate the code (include as many details as possible about the functionality).

Currently I submit 2 versions of vignettes:

- in [.pdf format to CRAN](https://cran.r-project.org/web/packages/IceSat2R/vignettes/IceSat-2_Atlas_products_PDF.pdf) where I use command ```system('R CMD build --compact-vignettes="gs+qpdf" IceSat2R')``` to reduce the .tar.gz file size
- in [.html format when I upload to Github](https://mlampros.github.io/IceSat2R/articles/IceSat-2_Atlas_products_HTML.html) (where visualizations can be displayed in the original true format)

And the current file size in in the limit of 5 MB
