### 2022-04-27 #4342 @mlampros
> And the software does not work as we will see later.
avail_cycles = available_RGTs(only_cycle_names = TRUE)
#> Error in seq.int(0, to0 - from, by) : 'to' must be a finite integer

I don't receive any errors when I run this line of code on my PC (Linux Mint), but I always encourage R users to open an issue to the [Github repository](https://github.com/mlampros/IceSat2R/issues) (I mention that also in the [blog post](http://mlampros.github.io/2022/02/12/IceSat2R_Altimetry_data/) that I always upload for a new R package)
