### 2022-04-27 #4342 @mlampros
>> Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

> No. The paper contains 3 references including a reference to the package itself (so 2 references). I'm not in ICESat stuff but this definitively looks light. There is no other API in other languages/software for ICEsat data?

In my submitted paper I only used 3 sources (including the R package). The R package itself cites [all authors for all IceSat-2 Products](https://cran.r-project.org/web/packages/IceSat2R/citation.html)
