### 2022-04-27 #4342 @mlampros
> Overall the code presented in the article is tedious to read and understand but could be very simple and smooth by vectorizing the functions and by doing all the pre-processing job internally. In R we are expecting all of this to be done by the package not by the user. This is why I said earlier that the package is not feature-complete.

High level functions can be included. The main reason that the current code uses for-loops is for debugging purposes, especially when downloading multiple IceSat-2 products or for instance when comparing the tracks between the OpenAltimetry API and the NSIDC service specifications.
