### 2022-04-27 #4342 @mlampros
>  I see in the README there is a shiny app but can't find easily how to open it.

The shiny application allows the user to create a global grid (or a grid based on an Area of Interest). The [select_aoi_global_grid](https://mlampros.github.io/IceSat2R/reference/select_aoi_global_grid.html) R6 class includes details in the example section how an IceSat2R user can take advantage of this shiny application to create either a 1- or 5-degree grid to perform the OpenAltimetry API queries (the OpenAltimetry API restricts the queries to a maximum of 1 or 5 degrees)
