### 2022-04-27 #4342 @Jean-Romain
I don't know if I'm supposed to answer every single argument. I won't but I think I can clarify some points. Actually I know why  you did what you did (CRAN policy, package size, check failures and so on) but I've been ask to say if yes/no the package fulfills a list of requirements. I looked into the packages and I answered factually if the answer is yes or no.

- I've been ask if the software contains automatic test. The answer is no. You can argue why (and I already knew why actually) but the fact is that there is no automatic testing. You can keep the vignette hard coded and keep the example wrapped in `\dontrun`. I do agree for the vignette I'm sometime doing the same. I disagree with the documentation but I won't argue. But if you want me to answer yes to this question a significant part one of the code must be tested either in the documentation or in unit tests.

- I've been asked if the the software documented to a satisfactory level. I admit I did not read the entire documentation. I looked at the documentation of the very first functions encountered in your tutorial and I think it's fair. They were not documented, period. You can argue that I did not look to the correct functions because that ones are not important functions and so on. This is not receivable. If the functions are really *secondary functions* this can be documented for example by explicitly mentioning that the functions are *internal function exported to debugging purpose*. This is legit and informative. But if it is really what they are, don't use them at the very beginning of your tutorial. Don't use them at all in your tutorial.

- My argument about tedious examples and for loops is more opinion-based I admit. I've have been asked if the software meets R standards. In R we almost never write for loop. Internally sometime we need loop but it is relatively rare. But certainly not on the user side and certainly not in an beginner tutorial. Consequently I think it is fair to answer that R standards are not met. You argument about debugging is not receivable. If the software is still in a debugging state or if the beginner tutorial aims for debugging it means that your package is not ready for production or that the tutorial for beginners is not a tutorial for beginners but a unit test.

- I opened an issue. I'll test the package more in depth when it will be fixed
