### 2022-04-27 mlampros/IceSat2R#3 @Jean-Romain
Error in seq.int(0, to0 - from, by) : 'to' must be a finite integer
I cannot run a single example of the documentation/vignette. `available_RGTs` gives the following error.

```r
avail_cycles = available_RGTs(only_cycle_names = TRUE)
#> Error in seq.int(0, to0 - from, by) : 'to' must be a finite integer
```

- Linux Mint
- R 4.1.3
- IceSat2R 1.0.1
