### 2022-04-27 #4342 @mlampros
> High level functions can be included. The main reason that the current code uses for-loops is for debugging purposes, especially when downloading multiple IceSat-2 products or for instance when comparing the tracks between the OpenAltimetry API and the NSIDC service specifications.

> Your argument about debugging is not receivable. If the software is still in a debugging state or if the beginner tutorial aims for debugging it means either that your package is not ready for production or that the tutorial for beginners is not a tutorial for beginners but a unit test.

I don't try to find excuses regarding the for-loops that I use. This is the way I write R code when I believe the code snippets that I use might give errors. The fact that I use for-loops in specific places in the vignettes and in the example sections is because I observed (in the past months) that in some cases, the calls to the OpenAltimetry API might not return the results instantly. This might had to do with the internet connection or with to many requests at the same time (I wasn't sure)
