### 2022-04-27 mlampros/IceSat2R#3 @mlampros
this is what I receive when I run the following code

```R
> require(IceSat2R)
Loading required package: IceSat2R
> avail_cycles = available_RGTs(only_cycle_names = TRUE)
> avail_cycles
 [1] "RGT_cycle_1"  "RGT_cycle_2"  "RGT_cycle_3"  "RGT_cycle_4"  "RGT_cycle_5"
 [6] "RGT_cycle_6"  "RGT_cycle_7"  "RGT_cycle_8"  "RGT_cycle_9"  "RGT_cycle_10"
[11] "RGT_cycle_11" "RGT_cycle_12" "RGT_cycle_13" "RGT_cycle_14" "RGT_cycle_15"
[16] "RGT_cycle_16"
```
on
- Linux Mint
- R 4.1.1
- IceSat2R version 1.0.1

I removed and re-installed the IceSat2R package from CRAN
