### 2022-04-27 mlampros/IceSat2R#3 @mlampros
I also run the previous code snippet on Rstudio Cloud with the following Operating Specifications
