### 2022-04-27 mlampros/IceSat2R#3 @Jean-Romain
The problem comme from the fact `avail_urls` contains NA

```
 1:              https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/antarcticaallorbits.zip                  Antarctic Orbit      <NA>    <NA>       Nominal
 2:                  https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/arcticallorbits.zip                     Arctic Orbit      <NA>    <NA>       Nominal
 3:    https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/ICESat2groundtracksWesternHem.zip Western Hemisphere Ground Tracks      <NA>    <NA>       Nominal
 4:    https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/ICESat2groundtracksEasternHem.zip Eastern Hemisphere Ground Tracks      <NA>    <NA>       Nominal
 5:        https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle1_date_time.zip        IS2_RGTs_cycle1_date_time      <NA>    <NA> Time_Specific
 6:        https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle2_date_time.zip        IS2_RGTs_cycle2_date_time      <NA>    <NA> Time_Specific
 7:      https://icesat-2.gsfc.nasa.gov/sites/default/files/page_files/IS2_RGTs_cycle3_date_time_0.zip      IS2_RGTs_cycle3_date_time_0      <NA>    <NA> Time_Specific
[...]
```

Then all dates are broken
