### 2022-04-27 mlampros/IceSat2R#3 @mlampros
For some reason you receive NA's for "Date_from" and "Date_to" whereas these 2 columns should have [either empty strings or dates ](https://github.com/mlampros/IceSat2R/blob/master/inst/data_files/technical_specs_urls.csv)
The [available_RGTs()](https://github.com/mlampros/IceSat2R/blob/master/R/mission_orbits.R#L399) function attempts first to download the latest metadata from the [technical_specs_url](https://icesat-2.gsfc.nasa.gov/science/specs) by using the [latest_orbits() ](https://github.com/mlampros/IceSat2R/blob/master/R/mission_orbits.R#L34)function (using internally the "rvest" R package). The NA's correspond by default to the "Nominal" and not to the "Time_specific" Type
