### 2022-04-27 mlampros/IceSat2R#3 @Jean-Romain
https://stackoverflow.com/questions/15566875/as-date-returning-na-while-converting-from-ddmmmyyyy

To put that in relationship to https://github.com/openjournals/joss-reviews/issues/4342:
- Pay attention to system locale.
- Write unit tests related to system locale.
- Read about system local (on my side I know nothing and I don't like dates)

```r
Sys.getlocale("LC_TIME")
#> [1] "fr_CA.UTF-8"
as.Date("October 13 2018", format = "%b %d %Y")
#> [[1] NA
lct <- Sys.getlocale("LC_TIME"); Sys.setlocale("LC_TIME", "C")
#> [[1] "C"
as.Date("October 13 2018", format = "%b %d %Y")
#> [[1] "2018-10-13
```
