### 2022-04-27 mlampros/IceSat2R#3 @mlampros
yes, it is a system locale issue, as per definition
```
%b
Abbreviated month name in the current locale on this platform. (Also matches
full name on input: in some locales there are no abbreviations of names.)
```
In my Operating System
```R
> dates_spec = list(c("October 13 2018", "December 28 2018"))
> dates_spec
[[1]]
[1] "October 13 2018"  "December 28 2018"

> dates_spec = lapply(dates_spec, function(x) {
+     as.character(as.Date(x, format = "%b %d %Y"))
+ })
> dates_spec
[[1]]
[1] "2018-10-13" "2018-12-28"
```
