### 2022-04-27 mlampros/IceSat2R#3 @mlampros
The only way to warn a user and avoid this issue is to throw an error
