### 2022-04-27 mlampros/IceSat2R#3 @Jean-Romain
I don't thing so. There are plenty of packages dealing with dates. You just need to learn how to handle the right way.
