### 2022-05-02 b16df1b
update R files and tests
