### 2022-05-02 mlampros/IceSat2R#3 @mlampros
@Jean-Romain,

- I've modified the code as I mention in the [NEWS.md file](https://github.com/mlampros/IceSat2R/blob/master/NEWS.md)
- I've also included tests for the cases where the functions download data. The user of the R package can test all files or specific ones as I [mention in the README.md file](https://github.com/mlampros/IceSat2R#r-package-tests)
- Regarding the system locale, I ended up modifying the code so that the dates are processed as character strings and the conversion does not take place.
