### 2022-05-02 #4342 @mlampros
I [updated the code of the IceSat2R](https://github.com/mlampros/IceSat2R/blob/master/NEWS.md#icesat2r-102) package. The majority of the examples take more than 5 seconds (which is prohibitive for CRAN - I tested it once again) therefore I include the *\dontrun*. I included more tests (as I mention also in the [opened issue](https://github.com/mlampros/IceSat2R/issues/3#issuecomment-1115114155)), however it's not possible to run these tests on CRAN due to time constraints and additional due to potential failures. I use in the majority of the newly added tests the [skip_on_cran()](https://cran.r-project.org/web/packages/testthat/vignettes/skipping.html) function. However, I updated the [README.md file with a section](https://github.com/mlampros/IceSat2R#r-package-tests) (where I include instructions) on how a user can run locally all or specific files of the tests so that potential errors can be fixed.
