### 2022-05-02 mlampros/IceSat2R#4 @Jean-Romain
1 output sublist did not have a valid 'description' and will be removed!
I tried again the first vignette with the correct locale. I got:

```r
rgt_winter = time_specific_orbits(date_from = start_date_w,
                                  date_to = end_date_w,
                                  RGT_cycle = NULL,
                                  download_method = 'curl',
                                  threads = parallel::detectCores(),
                                  verbose = TRUE)
#> The available Icesat-2 orbits will be red from 'https://icesat-2.gsfc.nasa.gov/science/specs' ...
#> Access the data of the technical specs website ...
#> Extract the .zip files and the corresponding titles ...
#> Keep the relevant data from the url's and titles ...
#> Process the nominal and time specific orbits separately ...
#> Adjust the Dates of the time specific orbits ...
#> Create the nominal orbits data.table ...
#> Create the time specific orbits data.table ...
#> Return a single data.table ...
#> Elapsed time: 0 hours and 0 minutes and 0 seconds.
#> ICESAT-2 orbits: 'Earliest-Date' is '2018-10-13'  'Latest-Date' is '2022-09-20'
#> -----------------------------------------------------
#> The .zip file of 'RGT_cycle_9' will be downloaded ...
#> -----------------------------------------------------
#>   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
#>                                  Dload  Upload   Total   Spent    Left  Speed
#> 100  140M  100  140M    0     0  1201k      0  0:01:59  0:01:59 --:--:-- 1340k
#> The downloaded .zip file will be extracted in the '/tmp/RtmphQC1V2/RGT_cycle_9' directory ...
#> Download and unzip the RGT_cycle_9 .zip file: Elapsed time: 0 hours and 2 minutes and 2 seconds.
#> 138 .kml files will be processed ...
#> Parallel processing of 138 .kml files using  4  threads starts ...
#> 1 output sublist did not have a valid 'description' and will be removed!
#> 1 output sublist did not have a valid 'description' and will be removed!
#> 1 output sublist did not have a valid 'description' and will be removed!
#> 1 output sublist did not have a valid 'description' and will be removed!
#> 1 output sublist did not have a valid 'description' and will be removed!
#>  138 times the same message
```

I got again and again and again the same message for each successive download. I got 824 times the same message for `RGT_cycle_9` and so on. At the end `rgt_winter` looks legit
