### 2022-05-02 mlampros/IceSat2R#4 @mlampros
I stumbled upon an error case during the [setup of binder](https://github.com/mlampros/IceSat2R#binder)
Although in my Operating System this message does not occur, in the binder image I observed that an additional sublist during the processing of the .kml files is included (this sublist does not have the same format as the majority). What are the dimensions of  the output "rgt_winter" object?
