### 2022-05-02 mlampros/IceSat2R#4 @Jean-Romain
It is correct it contains 91390 features and 14 fields
