### 2022-05-02 mlampros/IceSat2R#5 @Jean-Romain
Suggestion of simplifcation for the vignettes and article
As a user, as a beginner, as a person who do not known IceSat let me suggest you what kind of content I'd like to see in the vignette 1. This suggestion is related to my review where I said that the documentation and vignettes are tedious to read and overly complex. I suggest to create higher level functions and respect common data storage practice such as the vignette 1 looks like that.

```r
require(IceSat2R)
require(sf)
sf::sf_use_s2(use_s2 = FALSE)

data(greenl_sh_east)
bb = st_bbox(greenl_sh_east)

rgt_winter = time_specific_orbits("2020-12-15","2021-02-15", bbox = bb, threads = 4, verbose = TRUE)
rgt_summer = time_specific_orbits("2021-06-15","2021-08-15", bbox = bb, threads = 4, verbose = TRUE)
```

7 lines and we are already somewhere between line 35 and 45.

Then there is the "unique RGT" stuff. You may keep it as is or include a function in your package to do it. It looks standard and repetitive so it should go in the package in my opinion. No need to print stuff. Beginner are able to use `length(unq_rgt_winter)` themselves. 2 lines instead of 15

```r
unq_rgt_winter = clean_duplicates(rgt_subs_winter)
unq_rgt_summer = clean_duplicates(rgt_subs_summer)
```

Then we have the definitionof `verify_RGTs`. This  should be a function of the package. We gain another 40 lines of code.

Then we have twice a complex set of code around `verify_RGTs`. It becomes 2 lines of code because `verify_RGTs` already set the correct column names and is internally able to deal with standard data.

```r
ver_trc_winter = verify_RGTs(rgt_winter)
ver_trc_summer = verify_RGTs(rgt_summer)
```

Then we have

```r
greenl_grid = degrees_to_global_grid(greenl_sh_east,  degrees = 4.5, verbose = TRUE)
```

And I stopped here. The first half of the vignette 1 can be contained in 12 lines with a careful design. Nobody want to go into lines and lines and lines of tedious code in the first learning step :wink:
