### 2022-05-02 mlampros/IceSat2R#5 @mlampros
Thanks for reminding me the high level functions. I'll implement these in the next days and I'll notify you so that you can review the changes.
