+++ inst/joss_paper/paper.Rmd
@@ -85,15 +85,14 @@ knitr::include_graphics(file.path("figures", "himalayas_aoi.png"))
 
 # required R packages
 
-pkgs = c('IceSat2R', 'magrittr', 'sf', 'mapview', 'leaflet', 'knitr', 'glue',
-         'data.table', 'CopernicusDEM', 'terra', 'geodist', 'ggplot2')
+pkgs = c('IceSat2R', 'magrittr', 'sf', 'mapview', 'leaflet', 
+         'glue', 'data.table')
 load_pkgs = lapply(pkgs, require, character.only = TRUE)
 
 # load the AOI
 
-himl_pth = system.file('data_files', 'vignette_data', 'himalayas.RDS', 
+himl_pth = system.file('data_files', 'vignette_data', 'himalayas.RDS',
                        package = "IceSat2R")
-
 geoms_himal = readRDS(himl_pth)
 
 sf_wkt = sf::st_geometry(subset(geoms_himal, area_size == 'big'))
@@ -117,11 +116,10 @@ for (iter in 1:8) {
 
 # Extract the unique Reference Ground Tracks (RGTs)
 
-lst_out = unlist(lst_out, recursive = F)
+lst_out = unlist(lst_out, recursive = FALSE)
 unq_rgts = as.vector(unique(unlist(lapply(lst_out, function(x) x$RGT))))
 unq_rgts
 # [1] "96"   "157"  "363"  "538"  "599"  "805"  "866"  "1041" "1308" "1247"
-
 ```
 
 In this use case, we are interested in ICESat-2 data for a specific time period (from *'2020-01-01'* to *'2021-01-01'* - 1-year's data). We'll make use of the *IceSat2R::vsi_time_specific_orbits_wkt()* function which queries all *15 ICESat-2 RGTs cycles* (as of March 2022) to come to the RGTs intersection for the specified 1-year time interval,
@@ -142,8 +140,8 @@ The next map shows the 18 different Date-Time matches for our defined 1-year tim
 
 ```{r, eval = F, echo = T}
 
-orbit_cy = mapview::mapview(orb_cyc_multi, legend = F)
-AOI_wkt = mapview::mapview(sf_wkt, legend = F)
+orbit_cy = mapview::mapview(orb_cyc_multi, legend = FALSE)
+AOI_wkt = mapview::mapview(sf_wkt, legend = FALSE)
 
 lft = orbit_cy + AOI_wkt
 
@@ -162,39 +160,10 @@ The output of *'vsi_time_specific_orbits_wkt()'* can be verified with the *OpenA
 ```{r, eval = F, echo = T}
 
 bbx_aoi = sf::st_bbox(obj = sf_wkt)
-rows_match = rep(FALSE, nrow(orb_cyc_multi))
-
-for (row in 1:nrow(orb_cyc_multi)) {
-  
-  dat_item = orb_cyc_multi[row, , drop = F]
-  Date = as.Date(dat_item$Date_time)
-  
-  op_tra = IceSat2R::getTracks(minx = as.numeric(bbx_aoi['xmin']),
-                               miny = as.numeric(bbx_aoi['ymin']),
-                               maxx = as.numeric(bbx_aoi['xmax']),
-                               maxy = as.numeric(bbx_aoi['ymax']),
-                               date = as.character(Date),
-                               outputFormat = 'csv',
-                               download_method = 'curl',
-                               verbose = FALSE)
-  
-  date_obj = dat_item$Date_time
-  tim_rgt = glue::glue("Date: {date_obj} Time specific RGT: '{dat_item$RGT}'")
-  
-  if (nrow(op_tra) > 0) {
[…]
