+++ inst/joss_paper/paper.md
@@ -91,15 +91,14 @@ The following map shows the bounding box areas (small on the left and big on the
 ```r
 # required R packages
 
-pkgs = c('IceSat2R', 'magrittr', 'sf', 'mapview', 'leaflet', 'knitr', 'glue',
-         'data.table', 'CopernicusDEM', 'terra', 'geodist', 'ggplot2')
+pkgs = c('IceSat2R', 'magrittr', 'sf', 'mapview', 'leaflet', 
+         'glue', 'data.table')
 load_pkgs = lapply(pkgs, require, character.only = TRUE)
 
 # load the AOI
 
-himl_pth = system.file('data_files', 'vignette_data', 'himalayas.RDS', 
+himl_pth = system.file('data_files', 'vignette_data', 'himalayas.RDS',
                        package = "IceSat2R")
-
 geoms_himal = readRDS(himl_pth)
 
 sf_wkt = sf::st_geometry(subset(geoms_himal, area_size == 'big'))
@@ -123,7 +122,7 @@ for (iter in 1:8) {
 
 # Extract the unique Reference Ground Tracks (RGTs)
 
-lst_out = unlist(lst_out, recursive = F)
+lst_out = unlist(lst_out, recursive = FALSE)
 unq_rgts = as.vector(unique(unlist(lapply(lst_out, function(x) x$RGT))))
 unq_rgts
 # [1] "96"   "157"  "363"  "538"  "599"  "805"  "866"  "1041" "1308" "1247"
@@ -147,8 +146,8 @@ The next map shows the 18 different Date-Time matches for our defined 1-year tim
 
 
 ```r
-orbit_cy = mapview::mapview(orb_cyc_multi, legend = F)
-AOI_wkt = mapview::mapview(sf_wkt, legend = F)
+orbit_cy = mapview::mapview(orb_cyc_multi, legend = FALSE)
+AOI_wkt = mapview::mapview(sf_wkt, legend = FALSE)
 
 lft = orbit_cy + AOI_wkt
 
@@ -171,46 +170,17 @@ The output of *'vsi_time_specific_orbits_wkt()'* can be verified with the *OpenA
 
 ```r
 bbx_aoi = sf::st_bbox(obj = sf_wkt)
-rows_match = rep(FALSE, nrow(orb_cyc_multi))
-
-for (row in 1:nrow(orb_cyc_multi)) {
-  
-  dat_item = orb_cyc_multi[row, , drop = F]
-  Date = as.Date(dat_item$Date_time)
-  
-  op_tra = IceSat2R::getTracks(minx = as.numeric(bbx_aoi['xmin']),
-                               miny = as.numeric(bbx_aoi['ymin']),
-                               maxx = as.numeric(bbx_aoi['xmax']),
-                               maxy = as.numeric(bbx_aoi['ymax']),
-                               date = as.character(Date),
-                               outputFormat = 'csv',
-                               download_method = 'curl',
-                               verbose = FALSE)
-  
-  date_obj = dat_item$Date_time
-  tim_rgt = glue::glue("Date: {date_obj} Time specific RGT: '{dat_item$RGT}'")
-  
-  if (nrow(op_tra) > 0) {
-    iter_op_trac = paste(op_tra$track, collapse = ', ')
-    cat(glue::glue("{tim_rgt}  OpenAltimetry: '{iter_op_trac}'"), '\n')
-    
-    if (length(op_tra$track) == 1) {
-      if (op_tra$track == dat_item$RGT) {
-        rows_match[row] = TRUE
-      }
-    }
-  }
-  else {
-    cat(glue::glue("{tim_rgt} without an OpenAltimetry match!"), '\n')
-  }
-}
[…]
