### 2022-05-04 mlampros/IceSat2R#5 @mlampros
in the [latest commit](https://github.com/mlampros/IceSat2R/commit/b0354ae8da824665dc9b283b675948457f00a3cc),
- I added the *ne_10m_glaciated_areas.rda* and *RGT_cycle_14.rda* files in the *data* directory and I removed the *ne_10m_glaciated_areas.RDS* and *RGT_cycle_14.RDS* files from the *inst* directory
- I've added the *verify_RGTs()* function (in the *API_utils.R* file)
- I modified / updated the Vignettes, Example sections and the JOSS paper taking into account all previous mentioned changes

I mentioned all changes in the [NEWS.md](https://github.com/mlampros/IceSat2R/blob/master/NEWS.md#icesat2r-102) file. From the JOSS paper I removed the code related to the verification of the RGTs and I also simplified the last for-loop (a double for-loop is not required to show the functionality of the package). I also removed R package dependencies that are not used in the code.

I went through the code of the package once again and I performed the necessary changes. The vignettes (especially the "IceSat-2_Atlas_products.Rmd" file) include details and use cases. Although, it not required for an author of an R package to include a use case (or too many details) I'm used to explain the functionality of my R packages that way. Moreover, I expect that R users of my R packages have R programming rudimentary skills.
