### 2022-05-04 #4342 @mlampros
I've updated the code of the IceSat2R package based on the [new issue 5](https://github.com/mlampros/IceSat2R/issues/5#issuecomment-1117023790) in the Github repository.

One of these changes was related to my submitted paper. It would be wise to re-generate the paper .pdf file
