### 2022-05-04 c7cdf20
update the tic.R file
