+++ tic.R
@@ -1,5 +1,5 @@
 # installs dependencies, runs R CMD check, runs covr::codecov()
-do_package_checks()
+do_package_checks(build_args = c('--compact-vignettes="gs+qpdf"', '--resave-data'))
 
 if (ci_on_ghactions() && ci_has_env("BUILD_PKGDOWN")) {
   # creates pkgdown site and pushes to gh-pages branch
