### 2022-05-12 mlampros/IceSat2R#6 @evetion
Suggestion `chmod -R 777 /home/YOUR_DIR`
In your documentation you suggest to `chmod -R 777 /home/YOUR_DIR`, but this is a bad practice, as it opens up your (private) homedir to everyone (including write permissions) on your computer. I'm not sure what issues you ran into, but it's better to refer to the Docker manual (https://docs.docker.com/engine/install/linux-postinstall/#manage-docker-as-a-non-root-user) for such things.
