### 2022-05-12 mlampros/IceSat2R#8 @evetion
Consistent naming, consider changing from IceSat2 to ICESat-2
Your package is called `IceSat2R`, but the official name of the mission is `ICESat-2`. Note the capitals and dash. I'm not sure whether the package can be renamed (`ICESat2R?`), but I think the documentation could use consistent naming of ICESat-2 in the titles.
