### 2022-05-12 mlampros/IceSat2R#9 @evetion
Bug in `time_specific_orbits`: `Class attribute on column 3 of item 96 does not match with column 3 of item 1.`
From the documentation [here](https://mlampros.github.io/IceSat2R/articles/IceSat-2_Atlas_products_HTML.html), I ran

```
> start_date_w = "2020-12-15"
> end_date_w = "2021-02-15"
>
> rgt_winter = time_specific_orbits(date_from = start_date_w,
+                                   date_to = end_date_w,
+                                   RGT_cycle = NULL,
+                                   download_method = 'curl',
+                                   threads = parallel::detectCores(),
+                                   verbose = TRUE)
The available Icesat-2 orbits will be red from 'https://icesat-2.gsfc.nasa.gov/science/specs' ...
Access the data of the technical specs website ...
Extract the .zip files and the corresponding titles ...
Keep the relevant data from the url's and titles ...
Process the nominal and time specific orbits separately ...
Adjust the Dates of the time specific orbits ...
Create the nominal orbits data.table ...
Create the time specific orbits data.table ...
Return a single data.table ...
Elapsed time: 0 hours and 0 minutes and 0 seconds.
ICESAT-2 orbits: 'Earliest-Date' is '2018-10-13'  'Latest-Date' is '2022-09-20'
-----------------------------------------------------
The .zip file of 'RGT_cycle_9' will be downloaded ...
-----------------------------------------------------
                                 Dload  Upload   Total   Spent    Left  Speed
100  140M  100  140M    0     0  5630k      0  0:00:25  0:00:25 --:--:-- 2878k
The downloaded .zip file will be extracted in the '/var/folders/w_/152mmjvx4cx7y55ky886tp1h0000gn/T//RtmpukLAeZ/RGT_cycle_9' directory ...
Download and unzip the RGT_cycle_9 .zip file: Elapsed time: 0 hours and 0 minutes and 27 seconds.
138 .kml files will be processed ...
Parallel processing of 138 .kml files using  10  threads starts ...
Error in data.table::rbindlist(inner_obj) :
  Class attribute on column 3 of item 96 does not match with column 3 of item 1.
```
