### 2022-05-12 #4342 @evetion
I've reviewed the paper and documentation and ran some of the examples. I'm not a R user (installed it just for this), but am familiar with ICESat-2 and its data distribution. Disclosure: I'm also the author of a ICESat-2 data handling package, [SpaceLiDAR.jl](https://github.com/evetion/SpaceLiDAR.jl) in Julia, interfacing with the NSIDC, not OpenAltimetry.

Overall, I recognize the need for packages like these, and the RGT download and handling is something I haven't seen automated before. I appreciate the effort put into the documentation, binder and Docker instructions. 👍🏻

## Checklist
Here are the items that I couldn't check yet.

> Functionality: Have the functional claims of the software been confirmed?

I ran into some bug (https://github.com/mlampros/IceSat2R/issues/9) on one of my first tries. I'll test some more after the weekend.

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

There's a statement of need in the paper, but it's missing from the documentation itself. You could just copy the relevant text.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I can't find the state of the field in the paper. There are certainly other ICESat-2 related packages in other ecosystems?

> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is to the point, but quite short and could use a bit of editing. For example, already the first sentence contains `and as of October 2021`, which doesn't seem to fit. The code section takes up a lot of space, much of it purely scripting around the package functions.

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

The reference list is a bit light, in the sense that there's only 2 external citations (not counting the self-citation here). You cite the ICESat-2 ATL06 product, but could cite the ICESat-2 mission [paper](http://www.sciencedirect.com/science/article/pii/S0034425716305089)? Or maybe other packages, the website from which you download the RGTs?

## Editing
I made several other reports based on issues found in the documentation. These are:

- https://github.com/mlampros/IceSat2R/issues/8
- https://github.com/mlampros/IceSat2R/issues/7
- https://github.com/mlampros/IceSat2R/issues/6
