### 2022-05-13 mlampros/IceSat2R#9 @mlampros
[RGT_cycle_9_sample_files.zip](https://github.com/mlampros/IceSat2R/files/8684247/RGT_cycle_9_sample_files.zip)

The easiest way to debug this error is to run specific lines "time_specific_orbits" function and observe which sublist does not have the same dimensions compared to the remaining (the attached .zip file includes 2 .kml files for RGT cycle 9),

```R

setwd('directory where the sample .kml files are saved')
vec_files = c('IS2_RGT_0001_cycle9_24-Sep-2020.kml', 'IS2_RGT_0002_cycle9_24-Sep-2020.kml')
if (!any(c(file.exists(vec_files[1]), file.exists(vec_files[2])))) stop("The saved .kml files are not in the correct directory!")

require(doParallel)
require(utils)
require(sf)
require(glue)
require(foreach)
require(data.table)

threads = 4
LEN_subset = 2
verbose = TRUE

if (threads > 1 & LEN_subset > 1) {
  if (verbose) cat(glue::glue("Parallel processing of {LEN_subset} .kml files using  {threads}  threads starts ..."), '\n')

  if (.Platform$OS.type == "unix") {
    doParallel::registerDoParallel(cores = threads)
  }

  if (.Platform$OS.type == "windows") {
    cl = parallel::makePSOCKcluster(threads)
    doParallel::registerDoParallel(cl = cl)
  }
}

sf_objs = list()

for (idx in 1:LEN_subset) {
  if (verbose) {
    message("FILE: ", idx, "/", LEN_subset, "\r", appendLF = FALSE)
    utils::flush.console()
  }

  FILE = vec_files[idx]
  iter_file = vec_files[idx]
  lrs = sf::st_layers(dsn = iter_file)
  NAMS = lrs$name
  if (length(NAMS) == 0) stop(glue::glue("The file '{FILE}' does not include any layers!"), call. = F)
  nam_file = gsub('.kml', '', FILE)
  LEN_nams = length(NAMS)

  if (threads > 1 & LEN_subset > 1) {
    inner_obj = foreach::foreach(i = 1:LEN_nams) %dopar% {
      LAYER = NAMS[i]
      lr_dat = sf::st_read(dsn = iter_file, layer = LAYER, quiet = T)         # I expect each .kml file to consist of a single 'sfc_POINT'
      lr_dat
    }
  }
  else {
    inner_obj = lapply(1:LEN_nams, function(x) {
      LAYER = NAMS[x]
      lr_dat = sf::st_read(dsn = iter_file, layer = LAYER, quiet = T)         # I expect each .kml file to consist of a single 'sfc_POINT'
      lr_dat
    })
  }

  # class_obj = as.vector(unlist(lapply(inner_obj, function(x) class(sf::st_geometry(x))[1])))                      # I expect each sublist to be of type "sfc_POINT" and normally observations which are not (and can be for instance "sfc_LINESTRING") won't have a description column either. Therefore use the next line for removal
  descr_not_idx = which(as.vector(unlist(lapply(inner_obj, function(x) is.na(x$description) | x$description == ""))))
  LEN_exc = length(descr_not_idx)
  if (LEN_exc > 0) {
    if (verbose) message(glue::glue("{LEN_exc} output sublist did not have a valid 'description' and will be removed!"))
    inner_obj = inner_obj[-descr_not_idx]
  }

  inner_obj = sf::st_as_sf(data.table::rbindlist(inner_obj))
  sf_objs[[glue::glue("{nam_file}")]] = inner_obj
}

str(sf_objs)

# List of 2
# $ IS2_RGT_0001_cycle9_24-Sep-2020:Classes ‘sf’, ‘data.table’ and 'data.frame':	95 obs. of  12 variables:
#   ..$ Name        : chr [1:95] "" "" "" "" ...
# ..$ description : chr [1:95] "RGT 1\n24-Sep-2020 05:39:06\nDOY 268\nCycle 9" "RGT 1\n24-Sep-2020 05:40:06\nDOY 268\nCycle 9" "RGT 1\n24-Sep-2020 05:41:06\nDOY 268\nCycle 9" "RGT 1\n24-Sep-2020 05:42:06\nDOY 268\nCycle 9" ...
# ..$ timestamp   : POSIXct[1:95], format: NA NA NA NA ...
# ..$ begin       : POSIXct[1:95], format: NA NA NA NA ...
# ..$ end         : POSIXct[1:95], format: NA NA NA NA ...
# ..$ altitudeMode: chr [1:95] "clampToGround" "clampToGround" "clampToGround" "clampToGround" ...
# ..$ tessellate  : int [1:95] -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 ...
# ..$ extrude     : int [1:95] 0 0 0 0 0 0 0 0 0 0 ...
# ..$ visibility  : int [1:95] 1 1 1 1 1 1 1 1 1 1 ...
# ..$ drawOrder   : int [1:95] NA NA NA NA NA NA NA NA NA NA ...
# ..$ icon        : chr [1:95] NA NA NA NA ...
# ..$ geometry    :sfc_POINT of length 95; first list element:  'XYZ' num [1:3] -0.132 0.028 1
# ..- attr(*, "sf_column")= chr "geometry"
# ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
# .. ..- attr(*, "names")= chr [1:11] "Name" "description" "timestamp" "begin" ...
# $ IS2_RGT_0002_cycle9_24-Sep-2020:Classes ‘sf’, ‘data.table’ and 'data.frame':	95 obs. of  12 variables:
#   ..$ Name        : chr [1:95] "" "" "" "" ...
# ..$ description : chr [1:95] "RGT 2\n24-Sep-2020 07:13:23\nDOY 268\nCycle 9" "RGT 2\n24-Sep-2020 07:14:23\nDOY 268\nCycle 9" "RGT 2\n24-Sep-2020 07:15:23\nDOY 268\nCycle 9" "RGT 2\n24-Sep-2020 07:16:23\nDOY 268\nCycle 9" ...
# ..$ timestamp   : POSIXct[1:95], format: NA NA NA NA ...
# ..$ begin       : POSIXct[1:95], format: NA NA NA NA ...
# ..$ end         : POSIXct[1:95], format: NA NA NA NA ...
# ..$ altitudeMode: chr [1:95] "clampToGround" "clampToGround" "clampToGround" "clampToGround" ...
# ..$ tessellate  : int [1:95] -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 ...
# ..$ extrude     : int [1:95] 0 0 0 0 0 0 0 0 0 0 ...
# ..$ visibility  : int [1:95] 1 1 1 1 1 1 1 1 1 1 ...
# ..$ drawOrder   : int [1:95] NA NA NA NA NA NA NA NA NA NA ...
# ..$ icon        : chr [1:95] NA NA NA NA ...
# ..$ geometry    :sfc_POINT of length 95; first list element:  'XYZ' num [1:3] -23.74811 0.00707 1
# ..- attr(*, "sf_column")= chr "geometry"
# ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
# .. ..- attr(*, "names")= chr [1:11] "Name" "description" "timestamp" "begin" ...

sf_objs_dtbl = sf::st_as_sf(data.table::rbindlist(sf_objs))
sf_objs_dtbl

# Simple feature collection with 190 features and 11 fields
# Geometry type: POINT
# Dimension:     XYZ
# Bounding box:  xmin: -0.1318472 ymin: 0.02795893 xmax: -0.1318472 ymax: 0.02795893
# z_range:       zmin: 1 zmax: 1
# CRS:           4326
# First 10 features:
# Name                                   description timestamp begin  end  altitudeMode tessellate extrude visibility drawOrder icon                       geometry
# 1       RGT 1\n24-Sep-2020 05:39:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-0.1318472 0.02795...
# 2       RGT 1\n24-Sep-2020 05:40:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-0.5484491 3.86669...
# 3       RGT 1\n24-Sep-2020 05:41:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-0.9340453 7.70776...
# 4       RGT 1\n24-Sep-2020 05:42:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-1.322114 11.54864 1)
# 5       RGT 1\n24-Sep-2020 05:43:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA>  POINT Z (-1.71399 15.38886 1)
# 6       RGT 1\n24-Sep-2020 05:44:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-2.111149 19.22799 1)
# 7       RGT 1\n24-Sep-2020 05:45:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-2.515282 23.06563 1)
# 8       RGT 1\n24-Sep-2020 05:46:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA>  POINT Z (-2.928375 26.9014 1)
# 9       RGT 1\n24-Sep-2020 05:47:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-3.352826 30.73494 1)
# 10      RGT 1\n24-Sep-2020 05:48:06\nDOY 268\nCycle 9      <NA>  <NA> <NA> clampToGround         -1       0          1        NA <NA> POINT Z (-3.791596 34.56594 1)
```

Each sublist consists of a data.table with 95 rows and 12 columns and once I rbind these I receive a simple features object with 190 rows and 11 columns. Do you receive the same output objects for these 2 sample files?
