### 2022-05-13 mlampros/IceSat2R#8 @mlampros
Yes you might have right regarding the name of the R package. I used the same parameter names as is the case in OpenAltimetry but I didn't do the same for the package name. Now that I've already submitted the package to CRAN I think it won't be wise to rename the package due to the existing R users. I'll have a look to the documentation of the package once again in the weekend and I'll adjust the cases where the naming is not consistent.
