### 2022-05-13 mlampros/IceSat2R#6 @mlampros
Thanks for mentioning it. I updated the README.md file
