### 2022-05-13 mlampros/IceSat2R#7 @mlampros
I assume you mean this code snippet,

```R
#......................................................................
# compute also the difference in distance between the beam measurements
# and the DEM coordinates (based on the 30-meter resolution cells)
#......................................................................

merg_cells$dem_dif_dist = geodist::[geodist](https://rdrr.io/pkg/geodist/man/geodist.html)(x = merg_cells[, [c](https://rdrr.io/r/base/c.html)('longitude', 'latitude')],
                                           y = merg_cells[, [c](https://rdrr.io/r/base/c.html)('lon_dem30', 'lat_dem30')],
                                           paired = TRUE,
                                           measure = 'geodesic')

[summary](https://rdrr.io/r/base/summary.html)(merg_cells$dem_dif_dist)

 #   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
 # 0.2356  8.1619 11.5943 11.1490 14.3201 20.5394
```
