### 2022-05-13 mlampros/IceSat2R#7 @mlampros
> ICESat-2 is reference to the ellipsoid, while CopernicusDEM is referenced to the EGM2008 geoid

You are right on this. The [documentation page ](https://copernicus-dem-30m.s3.amazonaws.com/readme.html)mentions

> Data structure
> Each file is its own object in Amazon S3. The data are organised per tiles in original 1x1 degree grid.

and regarding the data [comparison between ICESat and Copernicus DEM](https://spacedata.copernicus.eu/documents/20126/0/GEO1988-CopernicusDEM-RP-001_ValidationReport_V1.0.pdf/edc88601-897e-82df-5604-b6b60cbecb57?t=1586527010843) (page 9 of 23) is mentioned,

> In  order  to  provide  a  most  accurate  data  comparison,  the  ICESat  elevation  values  have  been compared with the
> source DEM of the Copernicus DEM, the WorldDEM at its native resolution of 0.4’’ arc seconds. To compare the ICESat
> GLAS v33 reference points with the same horizontal, WGS84, and the vertical reference system, EGM08, as the Copernicus
> DEM/WorldDEM, a transformation has been applied.

I'll have a look into this over the weekend and I'll adjust the vignette accordingly.
Thanks for making me aware of this issue.
