+++ R/get_Tracks_data.R
@@ -36,7 +36,7 @@
 #' sf::sf_use_s2(use_s2 = FALSE)
 #'
 #' #..........................................................
-#' # In case we are interested to find the IceSat-2 tracks for
+#' # In case we are interested to find the ICESat-2 tracks for
 #' #     - a specific area (or bounding box)
 #' #     - a specific time interval (close to September 2021)
 #' #..........................................................
