+++ .github/workflows/icesat2_tech_specs.yaml
@@ -1,6 +1,6 @@
 
 #..............................................................
-# Update the IceSat-2 'Nominal' and 'Time Specific' Orbits, i.e
+# Update the ICESat-2 'Nominal' and 'Time Specific' Orbits, i.e
 # the 'technical_specs_urls.csv' file in the 'inst' directory
 #..............................................................
 
