### 2022-05-13 mlampros/IceSat2R#8 @mlampros
I [adjusted the name of ICESat-2](https://github.com/mlampros/IceSat2R/commit/2e213d934b0bf3f6ad6532b60ac27e83c7010894) in the files that was feasible
