### 2022-05-15 mlampros/IceSat2R#7 @mlampros
I downloaded a sample Copernicus DEM image ("Copernicus_DSM_COG_10_N27_00_E061_00_DEM.tif") from the [AWS open data registry](https://registry.opendata.aws/copernicus-dem/) using the "CopernicusDEM" R package.
It seems that there is no vertical coordinate reference system (CRS) in the metadata of the .tif file using "gdalinfo". I tried using the 'sf' R package,

```R

sf::gdal_utils(util = 'info', source = "Copernicus_DSM_COG_10_N27_00_E061_00_DEM.tif", quiet = FALSE)

# Driver: GTiff/GeoTIFF
# Files: /home/lampros/DIR_SAVE_DEM/Copernicus_DSM_COG_10_N27_00_E061_00_DEM.tif
# Size is 3600, 3600
# Coordinate System is:
# GEOGCS["WGS 84",
#     DATUM["WGS_1984",
#         SPHEROID["WGS 84",6378137,298.257223563,
#             AUTHORITY["EPSG","7030"]],
#         AUTHORITY["EPSG","6326"]],
#     PRIMEM["Greenwich",0],
#     UNIT["degree",0.0174532925199433],
#     AUTHORITY["EPSG","4326"]]
# Origin = (60.999861111111109,28.000138888888888)
# Pixel Size = (0.000277777777778,-0.000277777777778)
# Metadata:
#   AREA_OR_POINT=Point
# Image Structure Metadata:
#   COMPRESSION=DEFLATE
#   INTERLEAVE=BAND
# Corner Coordinates:
# Upper Left  (  60.9998611,  28.0001389) ( 60d59'59.50"E, 28d 0' 0.50"N)
# Lower Left  (  60.9998611,  27.0001389) ( 60d59'59.50"E, 27d 0' 0.50"N)
# Upper Right (  61.9998611,  28.0001389) ( 61d59'59.50"E, 28d 0' 0.50"N)
# Lower Right (  61.9998611,  27.0001389) ( 61d59'59.50"E, 27d 0' 0.50"N)
# Center      (  61.4998611,  27.5001389) ( 61d29'59.50"E, 27d30' 0.50"N)
# Band 1 Block=1024x1024 Type=Float32, ColorInterp=Gray
#   Overviews: 1800x1800, 900x900, 450x450

```

and using the latest [gdal ubuntu docker image](https://github.com/OSGeo/gdal/tree/master/docker#small-osgeogdalubuntu-small-latest),

```R
docker run --rm -v /home:/home/ osgeo/gdal:ubuntu-small-latest gdalinfo $PWD/Copernicus_DSM_COG_10_N27_00_E061_00_DEM.tif

# Driver: GTiff/GeoTIFF
# Files: /home/lampros/Copernicus_DSM_COG_10_N27_00_E061_00_DEM.tif
# Size is 3600, 3600
# Coordinate System is:
#   GEOGCRS["WGS 84",
#           DATUM["World Geodetic System 1984",
#                 ELLIPSOID["WGS 84",6378137,298.257223563,
#                           LENGTHUNIT["metre",1]]],
#           PRIMEM["Greenwich",0,
#                  ANGLEUNIT["degree",0.0174532925199433]],
#           CS[ellipsoidal,2],
#           AXIS["geodetic latitude (Lat)",north,
#                ORDER[1],
#                ANGLEUNIT["degree",0.0174532925199433]],
#           AXIS["geodetic longitude (Lon)",east,
#                ORDER[2],
#                ANGLEUNIT["degree",0.0174532925199433]],
#           ID["EPSG",4326]]
# Data axis to CRS axis mapping: 2,1
# Origin = (60.999861111111109,28.000138888888888)
# Pixel Size = (0.000277777777778,-0.000277777777778)
# Metadata:
#   AREA_OR_POINT=Point
# Image Structure Metadata:
#   COMPRESSION=DEFLATE
# INTERLEAVE=BAND
# LAYOUT=COG
# PREDICTOR=3
# Corner Coordinates:
#   Upper Left  (  60.9998611,  28.0001389) ( 60d59'59.50"E, 28d 0' 0.50"N)
# Lower Left  (  60.9998611,  27.0001389) ( 60d59'59.50"E, 27d 0' 0.50"N)
# Upper Right (  61.9998611,  28.0001389) ( 61d59'59.50"E, 28d 0' 0.50"N)
# Lower Right (  61.9998611,  27.0001389) ( 61d59'59.50"E, 27d 0' 0.50"N)
# Center      (  61.4998611,  27.5001389) ( 61d29'59.50"E, 27d30' 0.50"N)
# Band 1 Block=1024x1024 Type=Float32, ColorInterp=Gray
#   Overviews: 1800x1800, 900x900, 450x450

```

I would expect that the "VERT_CS" tag is present in the "gdalinfo" output as mentioned in this [stackoverflow thread](https://gis.stackexchange.com/a/413000).
The maintainer of the Copernicus DEM on AWS ("sentinelhub") [confirmed in a related question](https://forum.sentinel-hub.com/t/copernicus-dem-data-available-on-aws/3027/10) that the vertical CRS is the "EGM2008".
