### 2022-05-17 mlampros/IceSat2R#7 @mlampros
I re-run the vignette to use the Copernicus DEM .tif files that correspond (as the ICESat-2 data) to the area of interest. Then I modified the code in the following way,

```R
#...........................................................................................................
# The Copernicus DEM has a "horizontal" CRS (Coordinate Reference System) of WGS84-G1150 (EPSG 4326)
# and a "vertical" CRS of EGM2008 (EPSG 3855) - geoid. Therefore, a transformation of the vertical
# CRS is required to match the ellipsoid CRS of the ICESat-2 data
# The Table 1 (page 10 of 37) of the following weblink includes more information:
# https://spacedata.copernicus.eu/documents/20126/0/GEO1988-CopernicusDEM-SPE-002_ProductHandbook_I1.00.pdf
# The 2.5 minute geoid grid .gtx file of EGM2008 was downloaded from the following "osgeo" url:
# http://download.osgeo.org/proj/vdatum/egm08_25/
#...........................................................................................................

pth_gtx = file.path(dem_dir, 'egm08_25.gtx')

download.file(url = 'http://download.osgeo.org/proj/vdatum/egm08_25/egm08_25.gtx',
              destfile = pth_gtx,
              method = 'curl')

pth_dem_transformed = file.path(dem_dir, 'copernicus_dem_egm2008_transformed.tif')

convrt = sf::gdal_utils(
  util = "warp",
  source = pth_egm,
  destination = pth_dem_transformed,
  options = c("-s_srs", glue::glue("+proj=longlat +datum=WGS84 +no_defs +geoidgrids={pth_gtx}"),
              "-t_srs", "+proj=longlat +datum=WGS84 +no_def"),
  quiet = FALSE)

#....................................
# load the transformed Copernicus DEM
#....................................

rst_crop_transformed = terra::rast(x = pth_dem_transformed)

#...............................................
# we also have to find the closest elevation
# value to the 'winter' and 'summer' coordinates
# using the raster resolution
#...............................................

ter_dtbl = data.table::as.data.table(x = rst_crop_transformed, xy = TRUE, cells = TRUE)
colnames(ter_dtbl) = c("cell", "lon_dem30", "lat_dem30", "dem30")
# length(unique(ter_dtbl$cell)) == nrow(ter_dtbl)

xy = as.matrix(sw_hq_merg_beams[, c('longitude', 'latitude')])
sw_cells = terra::cellFromXY(object = rst_crop_transformed, xy = xy)

sw_hq_merg_beams$cell = sw_cells
# length(unique(sw_hq_merg_beams$cell)) < nrow(sw_hq_merg_beams)

merg_cells = merge(x = sw_hq_merg_beams, y = ter_dtbl, by = 'cell')

#......................................................................
# compute also the difference in distance between the beam measurements
# and the DEM coordinates (based on the 30-meter resolution cells)
#......................................................................

merg_cells$dem_dif_dist = geodist::geodist(x = merg_cells[, c('longitude', 'latitude')],
                                           y = merg_cells[, c('lon_dem30', 'lat_dem30')],
                                           paired = TRUE,
                                           measure = 'geodesic')

summary(merg_cells$dem_dif_dist)

 #   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
 # 0.1228  5.9179  8.9506  9.1205 12.5481 17.1965

```

Therefore,

- I used the 2.5 min. grid .gtx file to perform the transformation using "gdalwarp"
- then I computed the cells for the transformed Copernicus DEM and the ICESat-2 data
- then I finally merged both based on the grid cells and I computed the difference using the lat long values (of the ICESat-2 and Copernicus DEM)

The difference is reduced from

```R

#   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# 0.2356  8.1619 11.5943 11.1490 14.3201 20.5394

```
to

```R
 #   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
 # 0.1228  5.9179  8.9506  9.1205 12.5481 17.1965

```

Worth mentioning is that the vertically transformed Copernicus DEM elevation is now higher than the ICESat-2 land-ice height (both in winter and summer)

```R

# before transformation (previous summary data)

# summary(merg_cells$h_li_winter - merg_cells$dem30)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# 51.24   53.95   54.32   54.27   54.65   55.94

# summary(merg_cells$h_li_summer - merg_cells$dem30)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# 53.02   54.36   54.67   54.68   54.99   56.34

# after the transformation (using the .gtx file and gdalwarp)

# summary(merg_cells$h_li_winter - merg_cells$dem30)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# -49.80  -46.53  -46.01  -46.22  -45.68  -44.15

# summary(merg_cells$h_li_summer - merg_cells$dem30)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# -48.21  -46.16  -45.73  -45.81  -45.37  -44.04

```

What I would expect is that the ICESat-2 land-ice-height data would have higher values compared to the Copernicus DEM
