### 2022-05-17 mlampros/IceSat2R#7 @mlampros
An alternative way to come to the .gtx file is by using [GeographicLib](https://geographiclib.sourceforge.io/C++/doc/geoid.html)
I ported this C++ library in R and saved the .gtx file for different [intervals per degree](https://github.com/geographiclib/geographiclib/blob/main/examples/GeoidToGTX.cpp#L43) which allows a better resolution for the output .gtx file, however I observed that it doesn't give different results compared to the existing 2.5 minute geoid grid [downloaded from "osgeo"](http://download.osgeo.org/proj/vdatum/egm08_25/)
