+++ inst/joss_paper/paper.Rmd
@@ -29,7 +29,7 @@ knitr::opts_chunk$set(echo = FALSE,
 
 # Summary
 
-Lidar technology (light detection and ranging) is integrated into autonomous cars, unmanned aerial vehicles (UAV), airplanes, and satellites and as of October 2021, it represents a growing market. Many satellite missions utilize lidar to observe the surface of the earth and one of these is ICESat-2, which was launched on 15th September 2018 with the primary goal to measure changes in glaciers and ice sheets [@smith2019]. ICESat-2 (the successor of ICESat) consists of 6 beams (3 pairs), where each pair is separated by 3 kilometers and each beam in the pair is also separated by 90 meters as the following image shows [@smith2019],
+Lidar technology (light detection and ranging) is integrated into autonomous cars, unmanned aerial vehicles (UAV), airplanes, and satellites and it represents a growing market. Many satellite missions utilize lidar to observe the surface of the earth and one of these is ICESat-2 [@MARKUS2017260], which was launched on 15th September 2018 with the primary goal to measure changes in glaciers and ice sheets [@smith2019]. ICESat-2 (the successor of ICESat) consists of 6 beams (3 pairs), where each pair is separated by 3 kilometers and each beam in the pair is also separated by 90 meters as the following image shows [@smith2019],
 
 ```{r icesat-beams, fig.pos='h', out.width = "100%", out.height = "30%", fig.cap = "the ICESat-2 ATLAS six-beam pattern", fig.alt="The image shows the ICESat-2 ATLAS six-beam pattern"}
 knitr::include_graphics(file.path("figures", "icesat_2_details.jpg"))
@@ -67,13 +67,24 @@ knitr::include_graphics(file.path("figures", "icesat_2_diagram.png"))
 ```
 
 
+# State of the field
+
+In R programming there isn't currently a CRAN package that allows a user to download ICESat-2 data for the majority of the data products. 
+
+* The [icesat2R](https://github.com/mqueinnec/icesat2R) R package [@queinnec2022] appears only on Github and based on the [DESCRIPTION file](https://github.com/mqueinnec/icesat2R/blob/master/DESCRIPTION#L3) it includes code that allows the "Downloading and processing of ICESat-2 data for forest analysis" (Atlas-03 and Atlas-08 Products). 
+* In Python programming the [icepyx](https://github.com/icesat2py/icepyx) [@icepyx] is both a software library and a community composed of ICESat-2 data users, developers, and the scientific community that work together to develop a shared library of resources that simplify the process of querying, obtaining, analyzing, and manipulating ICESat-2 datasets to enable scientific discovery.
+* In Julia programming the [SpaceLiDAR.jl](https://github.com/evetion/SpaceLiDAR.jl) [@Pronk_SpaceLiDAR_jl_2021] is a research package that currently supports the "GLAH14 v34", "ATL03 v4", "ATL08 v4" and "ATL12 v4" ICESat-2 data products.
+
+The main difference between "IceSat2R" and the other programming packages ("icesat2R", "icepyx", "SpaceLiDAR.jl") is that the "IceSat2R" package allows the users to download ICESat-2 data using the OpenAltimetry API whereas users of the remaining software download ICESat-2 data directly from NSIDC.
+
+
 # Example use case
 
 The example code of this section explains how an R user can utilize the [IceSat2R](https://cran.r-project.org/web/packages/IceSat2R/index.html) package to,
 
[…]
