+++ inst/joss_paper/paper.bib
@@ -32,3 +32,43 @@
 	url = {https://CRAN.R-project.org/package=IceSat2R}
 }
 
+@misc{queinnec2022,
+  author = {Martin Queinnec},
+  title = {Downloading and processing ICESat-2 data for forest analysis},
+  year = {2022},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/mqueinnec/icesat2R}},
+  commit = {a3e00c7d0aeb526faa440a3ddbcc6185e47eb751}
+}
+
+@misc{icepyx,
+author = {Scheick, Jessica and others},
+organization = {icesat2py},
+title = {{icepyx: Python} tools for obtaining and working with {ICESat-2} data},
+year = {2019--},
+url = "https://github.com/icesat2py/icepyx"
+}
+
+@software{Pronk_SpaceLiDAR_jl_2021,
+author = {Pronk, Maarten},
+month = {7},
+title = {{SpaceLiDAR.jl}},
+url = {https://github.com/evetion/SpaceLiDAR.jl},
+version = {1.0.0},
+year = {2021}
+}
+
+@article{MARKUS2017260,
+title = {The Ice, Cloud, and land Elevation Satellite-2 (ICESat-2): Science requirements, concept, and implementation},
+journal = {Remote Sensing of Environment},
+volume = {190},
+pages = {260-273},
+year = {2017},
+issn = {0034-4257},
+doi = {https://doi.org/10.1016/j.rse.2016.12.029},
+url = {https://www.sciencedirect.com/science/article/pii/S0034425716305089},
+author = {Thorsten Markus and Tom Neumann and Anthony Martino and Waleed Abdalati and Kelly Brunt and Beata Csatho and Sinead Farrell and Helen Fricker and Alex Gardner and David Harding and Michael Jasinski and Ron Kwok and Lori Magruder and Dan Lubin and Scott Luthcke and James Morison and Ross Nelson and Amy Neuenschwander and Stephen Palm and Sorin Popescu and CK Shum and Bob E. Schutz and Benjamin Smith and Yuekui Yang and Jay Zwally},
+keywords = {ICESat-2, Land ice, Sea ice, Vegetation, Climate change, Satellite mission},
[…]
