### 2022-05-19 #4342 @mlampros
In the [latest commit](https://github.com/mlampros/IceSat2R/commit/24b1a0dd98ad99e2c0cc7b917feafb0bdceb6fa3) I updated the code of the IceSat2R package to account for the checklist of @evetion,

> I ran into some bug (https://github.com/mlampros/IceSat2R/issues/9) on one of my first tries. I'll test some more after the weekend.

I've [added code to debug ](https://github.com/mlampros/IceSat2R/issues/9#issuecomment-1125668956)the potential bug

> There's a statement of need in the paper, but it's missing from the documentation itself. You could just copy the relevant text.

I've added this [missing part in the documentation](https://mlampros.github.io/IceSat2R/reference/IceSat2R.html) of the IceSat2R package

> I can't find the state of the field in the paper. There are certainly other ICESat-2 related packages in other ecosystems?

I've added the [state of the field](https://github.com/mlampros/IceSat2R/blob/master/inst/joss_paper/paper.Rmd#L70) to the Joss paper

> The paper is to the point, but quite short and could use a bit of editing. For example, already the first sentence contains and as of October 2021, which doesn't seem to fit. The code section takes up a lot of space, much of it purely scripting around the package functions.

I've updated the Joss paper to exclude irrelevant text and keep to the limit of a submission. I also reduced the code section, which is mainly used in the example use case

> The reference list is a bit light, in the sense that there's only 2 external citations (not counting the self-citation here). You cite the ICESat-2 ATL06 product, but could cite the ICESat-2 mission [paper](http://www.sciencedirect.com/science/article/pii/S0034425716305089)? Or maybe other packages, the website from which you download the RGTs?

I've included further citations (state of field, ICESat-2 NSIDC Authors)

Regarding the remaining issues opened by @evetion,

- I [adjusted the name of ICESat-2](https://github.com/mlampros/IceSat2R/issues/8#issuecomment-1126263093) in the files that was feasible
- I updated the vignette regarding the Copernicus DEM comparison. I applied a [transformation using gdalwarp](https://github.com/mlampros/IceSat2R/issues/7#issuecomment-1128759469) and I also used the [GeographicLib to verify the output](https://github.com/mlampros/IceSat2R/issues/7#issuecomment-1128769500)
- I [updated the README.md file](https://github.com/mlampros/IceSat2R/issues/6#issuecomment-1125677538) to account for the issue of the docker image permissions
