+++ README.md
@@ -237,43 +237,29 @@ Using a table to map the **strong** and **weak** beams (Reference: [sliderule-py
 
 <br>
 
-<center> <h1>ATLAS oriented forward</h1> </center>
+<center> <h1>ATLAS oriented forward (+x)</h1> </center>
 
-+------------------------+------------------------------+--------------------+
-| **ATLAS Spot Number**  | **Ground track Designation** | **Beam Strength**  |
-+------------------------+------------------------------+--------------------+
+| ATLAS Spot Number      | Ground track Designation     | Beam Strength      |
+|:----------------------:|:----------------------------:|:------------------:|
 | 1                      | gt3r                         | Strong             |
-+------------------------+------------------------------+--------------------+
 | 2                      | gt3l                         | Weak               |
-+------------------------+------------------------------+--------------------+
 | 3                      | gt2r                         | Strong             |
-+------------------------+------------------------------+--------------------+
 | 4                      | gt2l                         | Weak               |
-+------------------------+------------------------------+--------------------+
 | 5                      | gt1r                         | Strong             |
-+------------------------+------------------------------+--------------------+
 | 6                      | gt1l                         | Weak               |
-+------------------------+------------------------------+--------------------+
 
 <br>
 
-<center> <h1>ATLAS oriented backwards</h1> </center>
+<center> <h1>ATLAS oriented backwards (-x)</h1> </center>
 
-+------------------------+------------------------------+--------------------+
-| **ATLAS Spot Number**  | **Ground track Designation** | **Beam Strength**  |
-+------------------------+------------------------------+--------------------+
+| ATLAS Spot Number      | Ground track Designation     | Beam Strength      |
+|:----------------------:|:----------------------------:|:------------------:|
 | 1                      | gt3r                         | Weak               |
-+------------------------+------------------------------+--------------------+
 | 2                      | gt3l                         | Strong             |
-+------------------------+------------------------------+--------------------+
 | 3                      | gt2r                         | Weak               |
-+------------------------+------------------------------+--------------------+
 | 4                      | gt2l                         | Strong             |
-+------------------------+------------------------------+--------------------+
 | 5                      | gt1r                         | Weak               |
-+------------------------+------------------------------+--------------------+
 | 6                      | gt1l                         | Strong             |
-+------------------------+------------------------------+--------------------+
 
 <br>
 
