### 2022-06-04 mlampros/IceSat2R#9 @evetion
That's a lot of code to debug something simple? Anyway, with your examples, I get the same error. The error stems from item 96, which is a Line, instead of Point like the others:

```r
[[95]]
Simple feature collection with 1 feature and 2 fields
Geometry type: POINT
Dimension:     XYZ
Bounding box:  xmin: -23.63926 ymin: -1.081025 xmax: -23.63926 ymax: -1.081025
z_range:       zmin: 1 zmax: 1
Geodetic CRS:  WGS 84
  Name                                Description                       geometry
1      RGT 1 24-Sep-2020 07:13:06 DOY 268 Cycle 9 POINT Z (-23.63926 -1.08102...

[[96]]
Simple feature collection with 1 feature and 2 fields
Geometry type: LINESTRING
Dimension:     XYZ
Bounding box:  xmin: -179.9708 ymin: -88.0062 xmax: 179.9952 ymax: 88.00649
z_range:       zmin: 0 zmax: 0
Geodetic CRS:  WGS 84
   Name Description                       geometry
1 RGT 1             LINESTRING Z (-0.1318472 0....
```
