### 2022-06-04 #4342 @evetion
> Functionality: Have the functional claims of the software been confirmed?

I've updated my bug report, but the issue is blocking for most functionality as described in the vignettes.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

There are no automated tests on Github. I think it would be good to run at least the vignettes there, so any errors are catched beforehand.
