### 2022-06-05 mlampros/IceSat2R#9 @mlampros
can you install the latest version using

```R
remotes::install_github('mlampros/IceSat2R')
```

 and re-try your example. It seems to me that I had a similar issue during the binder configuration when one of the R objects (sublists) were not of the same type (but this was in the first version of the package)
