### 2022-06-05 b966239
Deploy from Github Actions build 2441919013 [ci skip]
Build URL: https://github.com/mlampros/IceSat2R/actions/runs/2441919013
Commit: 905fd4a42bff70ea3ae928846a2e2c1828846b15
