### 2022-06-05 mlampros/IceSat2R#7 @mlampros
>  the first difference is the horizontal distance of the closest matching elevation?

yes this is the difference between the point coordinates of the ICESat-2 and their corresponding Copernicus DEM data

> And the vertical differences are now negative instead of positive?

the negative differences are the differences in elevation between the ICESat-2 and the Copernicus DEM data

I applied the correction once based on 'gdalwarp' using the 'sf' R package (you can see the code in my previous message).

The attached "copernicus_dem_egm2008.tif.zip" includes the Copernicus DEM that I used as "pth_egm" in the previous mentioned code snippet (inside the "sf::gdal_utils()" function) and the "egm08_25.gtx" file can be downloaded from [http://download.osgeo.org/proj/vdatum/egm08_25/](http://download.osgeo.org/proj/vdatum/egm08_25/) in case you want to reproduce (or verify) the differences. I don't think that I do something wrong.

[copernicus_dem_egm2008.tif.zip](https://github.com/mlampros/IceSat2R/files/8839478/copernicus_dem_egm2008.tif.zip)
