### 2022-06-05 mlampros/IceSat2R#7 @evetion
> yes this is the difference between the point coordinates of the ICESat-2 and their corresponding Copernicus DEM data

So if you only changed the raster elevation values, and not the resolution or bounds, these horizontal differences should've stayed the same?

> I don't think that I do something wrong.

I couldn't spot any mistakes as well, but I would argue that the results, either a ~+50m or ~-50m median difference are suspect. Especially given that the egm2008 geoid is ~+47m at that spot in Greenland. And ICESat-2, in ideal conditions, is known to be accurate within centimeters.
