### 2022-06-05 mlampros/IceSat2R#9 @evetion
I got the same error with the latest version.
