### 2022-06-05 mlampros/IceSat2R#9 @mlampros
normally the next code snippet should have shown a message regarding the mismatch in dimensions,

```R
  descr_not_idx = which(as.vector(unlist(lapply(inner_obj, function(x) is.na(x$description) | x$description == ""))))
  LEN_exc = length(descr_not_idx)
  if (LEN_exc > 0) {
    if (verbose) message(glue::glue("{LEN_exc} output sublist did not have a valid 'description' and will be removed!"))
    inner_obj = inner_obj[-descr_not_idx]
  }
```
before the

```R
inner_obj = sf::st_as_sf(data.table::rbindlist(inner_obj))
```
which throws the error

Is it possible that you create a new variable for the sublist with index 96 (of "inner_obj") and show me what you receive in the console using,

```R
subs = inner_obj[[96]]
str(subs)
subs$description
```
I would expect in case of a "LINESTRING" and not a "POINT" sf object that it would be removed.
