### 2022-06-05 mlampros/IceSat2R#7 @mlampros
> So if you only changed the raster elevation values, and not the resolution or bounds, these horizontal differences should've stayed the same?

That's true. Let me see if I have to explicitly specify these parameters (resolution, extend and CRS) of the input (source) image in the "gdalwarp" function (if it makes any difference and there is no difference compared to the initial computation)
