### 2022-06-05 mlampros/IceSat2R#7 @mlampros
It seems this was the reason for the negative differences in elevation. After specifying the resolution, extent and CRS (the following code snippet shows the changes compared to the previous code of my message),

```R

pth_egm = file.path(dem_dir, 'copernicus_dem_egm2008.tif')
rst_dem = terra::rast(pth_egm)
# rst_dem

CRS_dem = terra::crs(rst_dem, proj = TRUE)
RES_dem = as.character(terra::res(rst_dem))
EXT_dem = as.vector(terra::ext(rst_dem))
EXT_dem = c(EXT_dem['xmin'], EXT_dem['ymin'], EXT_dem['xmax'], EXT_dem['ymax'])
EXT_dem = as.character(EXT_dem)

convrt = sf::gdal_utils(
  util = "warp",
  source = pth_egm,
  destination = pth_dem_transformed,
  options = c("-s_srs", glue::glue("{CRS_dem} +geoidgrids={pth_gtx}"),
              "-t_srs", CRS_dem,
              "-tr", RES_dem,
              "-te", EXT_dem),
  quiet = FALSE)

```

I received the same summary statistics for the differences in elevation (between ICESat-2 and Copernicus DEM) as previously (so there is no change in the horizontal differences),

```R
summary(merg_cells$dem_dif_dist)
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
 0.2356  8.1619 11.5943 11.1490 14.3201 20.5394
```
and now the differences (between the ICESat-2 and Copernicus DEM data) for the summer and winter elevation data is positive,

```R
summary(merg_cells$h_li_winter - merg_cells$dem30)
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
 0.6775  3.6997  4.1428  4.0206  4.4779  5.8357
summary(merg_cells$h_li_summer - merg_cells$dem30)
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
  2.562   4.090   4.451   4.431   4.788   6.147
```
which is something would someone expect ( higher values for the land-ice-height ICESat-2 data compared to the Copernicus DEM)

I'll make the changes also to the code in the vignette and I'll push the changes to the repository
