### 2022-06-09 #4342 @Jean-Romain
I do not have the issue reported by @evetion. Or I least I don't think I have it because it seems I went further than him. Yet I stopped the process before the end. The first few lines of the tutorials take me something like 30-45 minutes to run. It is not reasonable for a tutorial. I suspect the non fixed issue (https://github.com/mlampros/IceSat2R/issues/4) I reported to increase even more the computation time but I might be wrong.

More than one month later my review is the same on most points. Few unit tests were added but they are all skipped on CRAN which invalidate the usefulness of unit tests. I have never been able to go further than half of the first tutorial because it is extremely tedious, much of it purely scripting around the package functions as said by @evetion. This is because the package does not include any high level function. I opened and issue (https://github.com/mlampros/IceSat2R/issues/5) to make some suggestions/ideas on how to improve that but the latest version does not contains significant changes on this point. Nothing I spotted at least. Consequently I won't go through the tutorials because each requires 45 minutes of data downloading/preprocessing and then the tutorial are mainly lines and lines of scripting not directly related to features of the package.

The documentation however includes significant changes and all the functions I checked are now documented.
