### 2022-06-09 mlampros/IceSat2R#7 @evetion
Those differences indeed look very sensible, glad we got it fixed :)
