### 2022-06-09 mlampros/IceSat2R#9 @evetion
```r
> subs = inner_obj[[96]]
> subs
Simple feature collection with 1 feature and 2 fields
Geometry type: LINESTRING
Dimension:     XYZ
Bounding box:  xmin: -179.9708 ymin: -88.0062 xmax: 179.9952 ymax: 88.00649
z_range:       zmin: 0 zmax: 0
Geodetic CRS:  WGS 84
   Name Description                       geometry
1 RGT 1             LINESTRING Z (-0.1318472 0....
> str(subs)
Classes ‘sf’ and 'data.frame':	1 obs. of  3 variables:
 $ Name       : chr "RGT 1"
 $ Description: chr ""
 $ geometry   :sfc_LINESTRING of length 1; first list element:  'XYZ' num [1:5656, 1:3] -0.132 -0.138 -0.145 -0.151 -0.157 ...
 - attr(*, "sf_column")= chr "geometry"
 - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA
  ..- attr(*, "names")= chr [1:2] "Name" "Description"
> subs$description
NULL
```

So, on closer inspection, there's a case sensitivity bug. I'm not sure where this stems from (I'm not a R nor KML user), but this is actually what you want no?

```r
> subs$Description
[1] ""
```
