### 2022-06-10 mlampros/IceSat2R#3 @mlampros
I close this issue for now, as I fixed the error case by modifying the code. Feel free to re-open it in case the code does not work as expected.
