### 2022-06-10 mlampros/IceSat2R#4 @mlampros
it seems this was not an issue as it was related to the warning that was added during the binder configuration (one of the sublists does not have the same format as the remaining - LINESTRING rather than POINT). This issue is also related to #9

I close this issue for now. Feel free to re-open it in case the code does not work as expected.
