### 2022-06-10 mlampros/IceSat2R#5 @mlampros
I explained the reason why a high level is not wise in the [Joss review](https://github.com/openjournals/joss-reviews/issues/4342#issuecomment-1151263399).

 Potential R users of this R package are free to open a pull request in case such a high level function can be beneficial for their work or scientific study.

I close this issue for now. Feel free to re-open it in case the code does not work as expected.
