### 2022-06-10 mlampros/IceSat2R#9 @mlampros
I'm wondering why this happens. We use the same ICESat-2 data but once you upload the data the 'description' column name appears with an uppercase 'D' in your case (as 'Description').
