### 2022-06-10 mlampros/IceSat2R#9 @evetion
I'm on a M1 Mac, although my filesystem isn't case sensitive, this might've something to do with it? Normally you would catch such platform issues with CI.
