### 2022-06-10 #4342 @hugoledoux
@mlampros I have read all the reviews and discussions (and issues the reviewers opened).

Let me start by saying that JOSS has a formal peer-review process that is designed to *improve the quality of the software submitted*, and our reviewers become (knowledgeable) users, they are asked to consider the software both from the point-of-views of the code, and from the users.

I have to agree with @Jean-Romain: you should have high-level functions that perform key tasks in an easy way. The point that you were paid for developing those functions is not relevant here, JOSS is trying to improve your package so that it is user-friendly, that its docs is up to a certain standard, etc. By saying that ["Potential R users of this R package are free to open a pull request in case such a high level function can be beneficial for their work or scientific study."](https://github.com/mlampros/IceSat2R/issues/5#issuecomment-1152012087), you omit that @Jean-Romain is one such user! Also, I do not understand your explanations about the high-level functions being problematic because one part of the process could go wrong, doesn't R have "try/catch" or exceptions handling? You could handle the potential errors and return an error code if something goes wrong during the process for instance.

I have also read the issues of @evetion and evidently your package wasn't tested on different OSes. Now we're in a weird situation where the reviewers are helping you to debug your code and they have to test and help you fix all the errors. This can be done to some extent for 1-2 issues, but it's not the idea of a review. I expect your code to be debugged and tested on different OSes, ideally with CI of some sort. This would allow to catch such errors.

I am especially worried about the comment that it takes 30-45min to run a tutorial, I think better should be done here. Isn't it possible to come up with a simple tutorial that uses a subset of a dataset? As long as the reviewers cannot run all the tests, your submission cannot be accepted by JOSS.

The good news is that JOSS is here to help you improve your submission, and we can wait for a while (we can put the review on pause). If you are willing to add: (1) higher levels that simply the life of users and are inline with what is considered normal by the R community; (2) unit tests that are debugged and tested by CI on different OSes; (3) better tutorials.

Let me know.
