### 2022-06-10 8bec407
updated NEWS.md and tests
