+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
-pkgdown: 2.0.3
+pkgdown: 2.0.4
 pkgdown_sha: ~
 articles:
   IceSat-2_Atlas_products_HTML: IceSat-2_Atlas_products_HTML.html
   IceSat-2_Mission_Orbits_HTML: IceSat-2_Mission_Orbits_HTML.html
   IceSat-2_Virtual_File_System_Orbits_HTML: IceSat-2_Virtual_File_System_Orbits_HTML.html
-last_built: 2022-06-10T04:59Z
+last_built: 2022-06-10T17:05Z
 
