+++ .github/workflows/tic.yml
@@ -120,6 +120,11 @@ jobs:
           echo -e 'options(Ncpus = 4, pkgType = "source", repos = structure(c(CRAN = "https://cloud.r-project.org/")))' > $HOME/.Rprofile
           Rscript -e "remotes::install_github('ropensci/tic')" -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
 
+      - name: "install the development version of the 'terra' package"
+        run: |
+          Rscript -e 'Sys.setenv("R_REMOTES_NO_ERRORS_FROM_WARNINGS" = "true")'
+          Rscript -e 'remotes::install_github("rspatial/terra")'
+
       - name: "[Stage] Script"
         run: Rscript -e 'tic::script()'
 
