+++ .github/workflows/tic.yml
@@ -122,8 +122,10 @@ jobs:
 
       - name: "install the development version of the 'terra' package"
         run: |
+          Rscript -e 'remove.packages(c("mapview", "terra"))'
           Rscript -e 'Sys.setenv("R_REMOTES_NO_ERRORS_FROM_WARNINGS" = "true")'
           Rscript -e 'remotes::install_github("rspatial/terra")'
+          Rscript -e 'remotes::install_github("r-spatial/mapview")'
 
       - name: "[Stage] Script"
         run: Rscript -e 'tic::script()'
