+++ .github/workflows/tic.yml
@@ -107,7 +107,10 @@ jobs:
         run: |
           # conflicts with gfortran from r-lib/actions when linking gcc
           rm '/usr/local/bin/gfortran'
-          brew install pkg-config gdal proj geos
+          # brew install pkg-config gdal proj geos
+          brew tap osgeo/osgeo4mac
+          brew install osgeo-gdal
+          brew install pkg-config proj geos
 
       - name: "[Stage] Install"
         if: matrix.config.os != 'macOS-latest' || matrix.config.r != 'devel'
@@ -119,11 +122,6 @@ jobs:
         run: |
           echo -e 'options(Ncpus = 4, pkgType = "source", repos = structure(c(CRAN = "https://cloud.r-project.org/")))' > $HOME/.Rprofile
           Rscript -e "remotes::install_github('ropensci/tic')" -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
-
-      - name: "install the development version of 'terra' and 'mapview' packages"
-        run: |
-          Rscript -e 'remotes::install_github("rspatial/terra", dependencies = TRUE)'
-          Rscript -e 'remotes::install_github("r-spatial/mapview", dependencies = TRUE)'
 
       - name: "[Stage] Script"
         run: Rscript -e 'tic::script()'
