+++ .github/workflows/tic.yml
@@ -107,10 +107,7 @@ jobs:
         run: |
           # conflicts with gfortran from r-lib/actions when linking gcc
           rm '/usr/local/bin/gfortran'
-          # brew install pkg-config gdal proj geos
-          brew tap osgeo/osgeo4mac
-          brew install osgeo-gdal
-          brew install pkg-config proj geos
+          brew install pkg-config gdal proj geos
 
       - name: "[Stage] Install"
         if: matrix.config.os != 'macOS-latest' || matrix.config.r != 'devel'
