### 2022-06-11 mlampros/IceSat2R#9 @mlampros
I've added a [test case](https://github.com/mlampros/IceSat2R/blob/master/tests/testthat/test-mission_orbits.R#L197-L212) to account for this issue on Mac osx, however the .tic configuration for Windows and Mac osx [fails since yesterday](https://github.com/mlampros/IceSat2R/actions/runs/2472745869) due to a "mapview" and "terra" configuration issue (it's not related to the added test case). I have to wait a few days to see if the issue resolves on its own (I tried to add a few dependencies but it didn't work)
