### 2022-06-14 #4342 @hugoledoux
I am disappointed to read this @mlampros.

The 3 requirements I describe above are what the reviewers and myself think is necessary to be accepted by JOSS.
