### 2022-06-14 #4342 @mlampros
@hugoledoux you shouldn't be disappointed.
Probably it wasn't a good decision to submit an API package to JOSS and also as an R package developer and maintainer to make the false assumption that this API package tutorial will be accepted.
As I said previously vignettes (or tutorials otherwise said) is not a requirement for an R package when submitting to CRAN. I just included these so that R users can have an idea of the various ways this API R package can be utilized.
My responsibility as an R package developer and maintainer is to fix potential issues (in the code and existing tutorials) and to have a look to submitted pull requests.
Feel free to close this review as from side I don't have to add something else. I'd like to thank again everyone spending time to review this API R package.
