### 2022-06-14 #4342 @Jean-Romain
Well, all of this is not required for CRAN acceptation this is correct. On my side I do not actually see the missing features as a problem for publication. I only answered to questions asked by JOSS. JOSS has publication requirements and I'm not questioning these requirements. JOSS wants unit tests? My answer is that there is not really any true unit tests. I did not conclude for rejection myself actually.

However, while I do agree it should not be a reason for rejection we have additional related problem here.

- There is no unit tests. I already made positive reviews in other journals for which this point was not an explicit requirement. But here we found 2 bugs preventing to use the package at the very beginning of the tutorials. No unit tests is always bad but when the packages works smoothly it is easier to relax some expectations. Especially CRAN run every example of the documentation so it already acts like some kind of unit tests. But in your package examples are wrapped in `\dontrun`. The documentation is not tested, no unit test and bugs that prevent to use the entire package. This is too much to argue that unit tests are not required.
- The package is designed in such a way that it is very hard to use it without a tutorial that explicitly tells users what to do and in which order to use the functions. With high level functions we could guess what to do from the doc and from the function names. But with low level functions that require a lot of scripting around, how the users are expected to figure out what to do? So, in my opinion the problem is not the absence of vignettes strictly speaking, The problem is that the vignettes are very complex and I was not able to go to the end and there is no way, in my opinion, to use the package without these vignettes. Yet you can argue that the package is designed for ICESAT users who already know ICESAT API and maybe for those users your functions make more sense. After all, the R `curl` API assumes that you already know `curl` for example. On my side I'm not convinced by this argument, I think it goes against what we are all trying to tend toward in R, i.e. providing super easy way to use complex stuff. But the argument is maybe receivable by JOSS.
