### 2022-06-14 #4342 @mlampros
> Especially CRAN run every example of the documentation so it already acts like some kind of unit tests. But in your package examples are wrapped in \dontrun. The documentation is not tested, no unit test and bugs that prevent to use the entire package. This is too much to argue that unit tests are not required.

I include tests (either for CRAN or only for Continuous Integration) that are feasible regarding execution time (https://github.com/mlampros/IceSat2R/tree/master/tests/testthat). The vignettes can not be tested because they take too long. I don't know if it would make sense to have only a single small tutorial which would consist of a few lines and would be feasible to run on CRAN or can be tested in Continuous Integration (Github Actions)

> The package is designed in such a way that it is very hard to use it without a tutorial that explicitly tells users what to do and in which order to use the functions

This is not correct. Please, have a look to the Example section of each function (https://mlampros.github.io/IceSat2R/reference/index.html) where I explain how the user should receive the results in detail (using all the functions required)
In this API package I included 3 vignettes with different level of complexity and length where the users can follow line by line to receive the results.

I think it was not a good idea to submit a version of one of my Vignettes to JOSS for review and I now realize that the [comment](https://github.com/openjournals/joss/issues/1057#issuecomment-1081060582) in my pre-submission inquiry (related to API's) was correct. I would prefer that editor closes this review.
