### 2022-06-15 #4342 @arfon
> I think it was not a good idea to submit a version of one of my Vignettes to JOSS for review and I now realize that the https://github.com/openjournals/joss/issues/1057#issuecomment-1081060582 in my pre-submission inquiry (related to API's) was correct. I would prefer that the editor closes this review.

Thanks for the productive conversation here folks. I'm happy to go ahead and withdraw the paper on behalf of the author.
